+++ src/include/scoring.h
@@ -362,7 +362,7 @@ auto get_top_k_from_scores(V const& scores, L&& top_k, size_t k = 0) {
 // size of the valid ranges in the scores matrix.
 // @todo pad top_k with sentinel if scores has sentinel
 template <class I, class T>
-auto get_top_k_from_scores(const ColMajorMatrix<T>& scores, int k_nn) {
+auto get_top_k_from_scores(const ColMajorMatrix<T>& scores, size_t k_nn) {
   auto top_k = ColMajorMatrix<I>(k_nn, scores.num_cols());
   for (size_t j = 0; j < scores.num_cols(); ++j) {
     get_top_k_from_scores(scores[j], top_k[j], k_nn);
